Speed-up some tests
Tests were executed against the parent commit of when they were
added to make sure that the tuning is not affecting the reproducibility
of the test. Note that some of the tests were sometimes not detecting
the bug, even when running with `-count=xxx`, so in that case, it
is difficult to know if the tuning would prevent detection if the
bug were to be reintroduced.

The FileStore and MsgTrace tests were also added as a matrix run
in Travis/GA.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>