When creating multiple workqueue consumers that collided, if the consumer landed on the meta-leader it would be created.
This was due to the fact that the meta-leader had the consumer assignment already and incorrectly passed wasExisting for isRecovering.
On recovery we would allow from a previous bug in the 2.9 series and print a warning to the server log.

The fix is to not mark waExisting if the consumer assignment is not marked as recovering, and to pass the state of isRecovering from the js state itself.

Also added upper layer check to avoid thrashing the meta layer.

If the parent stream was an R1 the consumer would be created, on R>1 it would be created and then a non meta-leader server would fail and clean it up.

Signed-off-by: Derek Collison <derek@nats.io>