[FIXED] Gateway: Outbound may fail to detect stale connection
If the first INFO never makes it back and for some reasons the
remote also does not close the connection, or connection is half
closed, the outbound would not have detected that the connection
is stale.

The outbound connection cannot send PING before it receives the
INFO from the remote because that is only when the outbound sends
the CONNECT. But the remote will fail if the first protocol from
an outbound is not CONNECT.

Using the same principle that we used for routes with compression
enabled, where PINGs cannot be sent until a certain time in the
handshake, we will use the ping timer to simply close the connection
as stale if the ping timer has not been reset before the time
needed to send the max number of pings.

Resolves #5349

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>