Improve leadership change signaling (#5504)
Because leadership status can sometimes change very quickly and
raft.updateLeadChange() can 'drop' some lead change signals, especially
with the leadc channel having a capacity of one, which means that
sometimes stream.setLeader(isLeader bool) gets called twice in a row
with the same isLeader value.

This could lead to in some cases the sourcing to 'stall' (not all
messages gets sourced) or 'duplicate messages' (more messages than
expected get sourced).

Changes are increasing the leadc channel's size to 32.

The effects of the change can be demonstrated by running the
TestStreamSourcingScalingSourcingManyBenchmark load test a large number
of times (e.g. 100 times): without these changes the test flaps and
fails between 10 and 20% of the time (most of the time 'stalling' but
also sometimes with 'too many'). With the changes over around a hundred
runs not once did the test fail with 'stalling' and only once did it
fail with 'too many' messages (so there's still a bug/corner case that
could result in 'more than expected' messages to be sourced, although it
seems to be a rare occurence (i.e. 1% of the time)).

This is a replacement for #5496

---------

Signed-off-by: Jean-Noël Moyne <jnmoyne@gmail.com>