Improvements to catchups and health checks.
These are mostly for memory based streams and cosnumers that need to rebuild complete state on a server restart.

For stream health checks for replicated streams make sure that the monitor routine is running.
When waiting on consumer assignments at the beginning of the stream monitor, make sure the consumer monitor is running as well if replicated.
On a consumer snapshot, register pre-acks as needed.
On stream checkInterestState reset an empty stream to the low ack floor from all consumers.

Signed-off-by: Derek Collison <derek@nats.io>