ProposeDirect would bypass internal prop ipq which could cause interleaving of state.
With skipMsgs calling ProposeDirect, if the prop queue was not empty this would place the skip entry ahead of the other entries.
This would result in sequence mismatch and stalled processing.

Bug was we should not bypass the internal prop ipq for the nrg. This corrects it and renames to ProposeMulti.

Signed-off-by: Derek Collison <derek@nats.io>