[FIXED] Subscription interest could remain after slow consumer error
When a slow consumer error is detected, the server will close the
connection. In some race situations, it could be that the subscription
interest is left intact, which would cause routed servers to continue
routing messages although the interest should have been gone.

The subscription could still be seen in the monitoring subscriptions
list (when displaying with details), referencing a client connection
that would be in the closed connection state.
The monitoring route page of the other servers would still indicate
that there is interest for this subject on the origin server.

Resolves #5539

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>