Fix 0 expected last seq per subject for clustered stream
The `getExpectedLastSeqPerSubject` check in `processJetStreamMsg` would take into account that we could get a create/seq=0 for last seq per subject.
However, the `processClusteredInboundMsg` would NOT take this into account. This would result in sending the message to all replicas. If any replica would store instead of reject the message, it would desync.

Signed-off-by: Maurice van Veen <github@mauricevanveen.com>