Make sure to not load blocks unnecessarily when checking if we can skip on initial miss in LoadNextMsg().
When we would miss a LoadNextMsg() from the block associated with the starting sequence, we would check if we could skip ahead.
The logic for this could load old blocks, behind where we were, if the psim fblk was outdated, causing memory bloat for cache expiration time intervals.

We reworked this to load no blocks in the skip check since this is hot path for LoadNextMsg() and the linear scan, if applicable, properly will expire anything loaded that did not contain a match.

Signed-off-by: Derek Collison <derek@nats.io>