NRG: Wait for goroutines on shutdown, hold JS lock on group create
Holding the JS lock means that it isn't possible for two calls to
`createRaftGroup` can't race with each other, resulting in multiple
instances of the same group running.

Also wait for the shutdown of all Raft-related goroutines so that a call
to `Stop()` or `Delete()` only returns once they have finished. Otherwise
a call to `Stop()` and `Delete()` can also race, resulting in the store
directory being prematurely deleted.

Signed-off-by: Neil Twigg <neil@nats.io>