Fix reflect.DeepEqual both receiving pointers
In `processClusterCreateStream` there was a call to `reflect.DeepEqual` with different types making it always fail.
`mset.config()` is the config itself, whereas `sa.Config` is a pointer to the config. So we need to pass `mset.config()` as a pointer instead.

In `processClusterCreateConsumer` this was done as well:
```go
cfg := o.config()
if isConfigUpdate = !reflect.DeepEqual(&cfg, ca.Config); isConfigUpdate {
```

Signed-off-by: Maurice van Veen <github@mauricevanveen.com>