NRG: Do not allow `InstallSnapshot` and `SendSnapshot` on empty snapshots
There's no reason we should be sending empty snapshots, and at the worst
they can be destructive to data. But since we may potentially be using
custom JSON or text marshallers, we need to be sensitive to the fact that
JSON marshal operations can potentially fail if something is misconfigured.

Signed-off-by: Neil Twigg <neil@nats.io>