[FIXED] Backport retry mset.ackMsg if removal fails (#6519)
Partial backport of #6140 for v2.10.26+

`mset.ackMsg` could fail if the clustered stream is behind on applies on
this server, but the consumer's ack floor is ahead. In this case
`checkStateForInterestStream` would skip its check floor ahead, never
retrying to ack/remove this message again. Which would leave messages
around, not being removed even though they could be.

This PR is a partial backport, still doing `mset.ackMsg` for each
individual server instead of via message delete proposals for clustered
streams, but allowing to retry if a removal should be done.

Signed-off-by: Maurice van Veen <github@mauricevanveen.com>

Signed-off-by: Maurice van Veen <github@mauricevanveen.com>
Signed-off-by: Neil Twigg <neil@nats.io>
Co-authored-by: Neil Twigg <neil@nats.io>