Stream leader should propose consumer remaps once assignment processed (#7071)
Right now, in a stream scale up/scale down scenario where there are
consumers that also need to be scaled, the metaleader makes a proposal
to update the stream and then immediately makes proposals to update the
consumers. However, this does not account for the fact that the stream
proposal _could_ fail to be applied, which is where we have sometimes
seen peer set drifts.

This moves this logic to take place on the stream leader once the stream
update proposal is applied instead.

This will also be necessary in a world where the consumer assignments
are managed by the stream leader, as the metaleader will not know
anything about the stream NRG.

Signed-off-by: Neil Twigg <neil@nats.io>