(2.12) [IMPROVED] Incremental filestore index.db
Currently, `index.db` needs to be written during a graceful shutdown. This means if you have a 100GB stream and the server gets hard killed without writing a new `index.db`. It will be invalidated, and the full stream will be scanned to rebuild the whole index.

This change proposes to make recovery based on `index.db` incremental. This means that a server after hard kill can recover with an older `index.db`, and require recovery only for a smaller part of the stream. If the last message block got more writes, those will be recovered, as well as any writes to later blocks, without needing to scan through the whole stream to fully rebuild.

However, if the `index.db` references a block that has been removed we need to remove its subject state. Similarly, we also need to do this for the `fs.lmb` when the `index.db` has a sequence that's not `fs.lmb.last`. The `index.db` doesn't contain subject state per block, nor do the remaining blocks on disk have a separate file that only contains subject state (I believe this is what `1.idx` files used to be for?). This means in that case we'll still need to go through the whole stream to recalculate and get the correct subject state.

We'll probably need to look into how we store subject state in an incremental way in general when picking up https://github.com/nats-io/nats-server/issues/6561. This change can be one part of the puzzle in improving how we index and restore after ungraceful restarts, but can't be used without solving the former.

Signed-off-by: Maurice van Veen <github@mauricevanveen.com>