[FIXED] Route: Infinite retry cluster connection with invalid credentials
With route pooling, the server would create a connection to the remote
and when receiving the INFO protocol, would start a new route if the
amount of routes was below the effective pool size. But if the route
failed due to authentication error, this would cause a rapid infinite
attempt to recreate a route.

The approach here is to delay the creation of the new route after
receiving the first PONG after sending a PING.

Resolves #7194

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>