Refactor processAppendEntry()
Extract recoverAppendEntry() from processAppendEntry(),
separating the normal path from recovery path. This allows
to simplify processAppendEntry() by removing checks for
for `sub == nil`, as subscription is always given

Extract processAppendEntryLocked() from processAppendEntry().
processAppendEntryLocked() does the actual processing of the
appendEntry under lock protection, and optionally returns a
pointer to a appendEntryResponse. processAppendEntry() is
responsible for sending the response and returning it to its
pool (if any).

Signed-off-by: Daniele Sciascia <daniele@nats.io>