Add ExportAccountConfig
This function derives account and user information from the `Options`
type.

The use case is to derived operator-mode account and user JWTs from
config-mode auth configuration, such as `users`, `accounts`, `mappings`,
etc. This would be done in an external tool.

Note, `Options` does not export sufficient information/fields to access
this information when embedding the server for tooling purposes.

Signed-off-by: Byron Ruth <byron@nats.io>