NRG: prevent peer remove of the last peer
Make sure that a Raft cluster cannot remove its last remaining peer.
Attempting to remove the last peer will now return an error, preventing
the cluster from entering an invalid state.

Test TestNRGProposeRemovePeerFollowers has been renamed to
TestNRGProposeRemovePeerFollowers and updated to cover this new
behavior, specifically testing the scenario where an attempt is made
to remove the leader when it is the only remaining peer.

Signed-off-by: Daniele Sciascia <daniele@nats.io>