NRG: Prevent disjoint majorities under peer additions
This commit fixes a bug where disjoint majorities could form
during peer additions. Previously, node would update their
cluster size only after the `EntryPeerAdd` log entry was
committed by the cluster. This led to a situation where a
leader could have a new peer in its peer set, before committing
the corresponding EntryPeerAdd, while keeping the quorum size
unchanged. This created the opportunity for a partitioned
leader to commit EntryPeerAdd using only append entry responses
from newly added nodes. For instance, a partitioned leader in
a three node cluster, could form a majority with a newly added
peer. Resulting in a 4 node membership, now requiring a quorum
of 3. A subsequent peer addition would also succeed, using
with the help of the two newly added nodes. Resulting in a
5 node membership with 3 functioning nodes on one side of the
partition, and the old followers forming a different majority
on the other side of the partition.

The fix ensures that a node adds a new peer and adjusts its
quorum calculation at the same time, right after appending the
corresponding EntryAddPeer to its WAL. This is in line with
the membership change algorithm given by Raft's original
author (section 4.1 of Ongaro's thesis).
With the revised logic, a partitioned leader of a 3 node cluster
is unable to form a majority using the acknowledgement from the
to be added peer. As the quorum now requires at least one
acknowledgement coming from the original followers.

Signed-off-by: Daniele Sciascia <daniele@nats.io>